EfficiencyTrack: Rebin histograms if there are more bins than events
If a certain ratio of bins/events is exceeded, the histograms for
alignment (and visual checking) are rebinned to contain less bins than
events. That makes fitting and therefore alignment possible.

Additionaly, the maxialy valued bin is used as initial point for fitting,
while the arithmetric mean between maximum bin and statistical mean
is used when the histogram had to be rebinned. Works well for most cases
but requires probably some adjustments or multiple tries.