StripEfficiency: Use fixed value to correlate tracks and strip activations
Before, 3*align.sigma was used for cutting, but sometimes StripAlign
produced unrealisticly low values for sigma. As a result, there are very
few correlated events and efficiency drops. Since the geometry is always
the same (and the rotational alignment seems to be consistently good), a
fixed, hard coded value is used ATM.